Renamed base.py to shell_info which is more relevant
