+++ tardis/widgets/__init__.py
@@ -1,2 +1,5 @@
-from tardis.widgets.base import shell_info_from_simulation, shell_info_from_hdf
+from tardis.widgets.shell_info import (
+    shell_info_from_simulation,
+    shell_info_from_hdf,
+)
 from tardis.widgets.line_info import LineInfoWidget
