+++ tardis/widgets/shell_info.py
